==> 01_Barbon_GasPrice_EDA.ipynb <==
# Manage libraries


import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

# Gas data

dfGas = pd.read_parquet("/Users/marcvendramet/Desktop/HSG/Master/MA/08_Data/gas_ts_block.par", engine='pyarrow')
dfGasCSV = dfGas.to_csv()

# Export to a csv file (around 700MB)
# dfGas.to_csv("/Users/marcvendramet/Desktop/HSG/Master/MA/08_Data/gas_ts_block_csv.csv")

## Exploration

## Findings
- index refers to block
- 12423140 blocks observed
- blocks start in 2015 (block: 46147)
- tx = number of transactions per block


dfGas.head()
dfGas.iloc[dfGas["avg_gas"].idxmax()]

type(dfGas)

dfGas.describe()

dfGas.index

len(dfGas)

max(dfGas["tx"])

max(dfGas["tot_gas"])

min(dfGas["tot_gas"])

dfGas.isna().any()


# takes too long to execute
# sns.histplot(dfGas['avg_gas'])

dfGas.hist()

plt.plot(dfGas.index, dfGas["avg_gas"])
plt.xscale("log")
