==> 02_CollectLP_TheGraph.ipynb <==
# Sources

* TWD approach with update(): https://towardsdatascience.com/exploring-decentraland-marketplace-sales-with-thegraph-and-graphql-2f5e8e7199b5
* Pagination of graphQL queries: https://stackoverflow.com/questions/59666982/how-to-use-variables-in-python-gql
* Pagination doc TheGraph: https://thegraph.com/docs/en/developer/graphql-api/


# Packages

import requests
import json
from gql import gql, Client # not installed in Thesis39 env
import pandas as pd
from gql.transport.requests import RequestsHTTPTransport

r = requests.get('https://api.github.com/events') # testing import requests

# Create dynamic API request

# Select your transport with a defined url endpoint
sample_transport = RequestsHTTPTransport(url='https://api.thegraph.com/subgraphs/name/uniswap/uniswap-v2',
                                         verify=True,
                                         retries=3,)
# Create a GraphQL client using the defined transport
client = Client(transport=sample_transport,
                fetch_schema_from_transport=True,
                )

myString = '''
query manyPairs($lastID: ID) {
    pairs(first: 1000, where: {id_gt: $lastID}) {
      id
     token0 {
      id
      symbol
    }
    token1 {
      id
      symbol
    }
    reserveUSD
    volumeUSD
   }
    }
'''

params = {
    "lastID": ""
}



df = pd.DataFrame(columns = ['poolId','token0','token1','volumeUSD','reserveUSD'])

# First run has no 'lastID'
params["lastID"] = ""

# Create loop with pagination on TheGraph

while True:
    queryAllLp = gql(myString)
    response = client.execute(queryAllLp, variable_values=params)

    #if there is no data returned it means you reached the end and should stop querying.
    if len(client.execute(queryAllLp, variable_values=params)['pairs']) <= 1:
        break

    else:
        df1 = []
        for i in response['pairs']:
            df1.append([
                i['id'],
                i['token0']['symbol'],
                i['token1']['symbol'],
                i['volumeUSD'],
                i['reserveUSD']
            ])
        df1 = pd.DataFrame(df1, columns = ['poolId','token0','token1','volumeUSD','reserveUSD'])

        # append your temp dataframe to your master dataset.
        df = df.append(df1)

        #renew counter variable
        params["lastID"] = df.iloc[-1,0]


# Save cleaned LPs as csv

# add column names to finished df after query

len(df["poolId"].unique())
df.head()
df.to_csv("/Users/marcvendramet/Desktop/HSG/Master/MA/08_Data/04_TheGraph/LP_UniswapV2_raw.csv")
