==> 03_LP cleaning.ipynb <==
# Packages

import pandas as pd
import numpy as np
#import seaborn as sns
#import matplotlib as plt
import re 


# Import file and exploration

df = pd.read_csv("/Volumes/Extreme SSD/03_TheGraph/LP_UniswapV2_raw.csv")

df.head()
df.tail()
df.shape
df.info()
df.dtypes
df.count()
df.describe()



# First Colum is repeating index values between 0-1000  from looping through the API. To be removed.
len(df.iloc[:,0].unique())

# Remove first column
dfClean = df.drop("Unnamed: 0", axis =1)

# Checking columns for NULL values
print(df.isnull().sum())
print(dfClean.isnull().sum())

# Removing Null values

# Null values for token0
null0 = df[df["token0"].isnull()]

# Null values for token1
null1 = df[df["token1"].isnull()]

# Cheeck position of example
dfClean.loc[40620]

# Create df with rows missing values
dfNull = null0.append(null1)

# Remove NULL observations
dfClean = dfClean.drop(dfNull.iloc[:,0].index)

# Check result table
print(dfClean.isnull().sum(), "\n")

print("Shape: ", dfClean.shape)

# Cleaning table

# Remove commas from value strings in columns that are object-type
for i in range(len(dfClean.columns)):
    if (dfClean.dtypes[i]) == "object":
        dfClean.iloc[:,i] = dfClean.iloc[:,i].replace(",", "", regex=True)
        dfClean.iloc[:,i] = dfClean.iloc[:,i].replace('"', "", regex=True)
        dfClean.iloc[:,i] = dfClean.iloc[:,i].replace(r'"', "", regex=True)
        dfClean.iloc[:,i] = dfClean.iloc[:,i].replace(r'<.+>', "", regex=True)
        dfClean.iloc[:,i] = dfClean.iloc[:,i].replace("<", "", regex=True)
        dfClean.iloc[:,i] = dfClean.iloc[:,i].replace(">", "", regex=True)


# Add flag column
dfClean["UniswapV2_LP"] = 1

# Add UniswapV2Router Address

#dfClean.loc[len(dfClean.index)] = ['test', 1] 
dfTemp = {'poolId': '0x7a250d5630b4cf539739df2c5dacb4c659f2488d', 'UniswapV2_LP': 1}
dfClean = dfClean.append(dfTemp, ignore_index = True)

# Export as csv



# Export all columns from cleaned df. Doesnt work for upload to BQ. Use '_export'
dfClean.to_csv("/Users/marcvendramet/Desktop/HSG/Master/MA/08_Data/04_TheGraph/LP_UniswapV2_clean.csv"

# Export only two columns for useabulity on BigQuery
dfExport = dfClean[["poolId", "UniswapV2_LP"]]

dfExport.to_csv("/Volumes/Extreme SSD/03_TheGraph/LP_UniswapV2_export.csv", index_label = "index", index = True)

## Archived: Labels not needed
clabels = dfClean.columns.tolist()
clabels.insert(0, "index")
print(clabels)
