==> 04_BGExport_v2.ipynb <==
# Useful Links
Useful links:
* Export job to GCS: https://blog.getcensus.com/3-ways-to-export-csv-files-from-google-bigquery/
* Export job to GCS: https://hevodata.com/learn/export-bigquery-table-to-csv/#met3
* Export job to GCS: https://cloud.google.com/bigquery/docs/samples/bigquery-extract-table#bigquery_extract_table-python
* Query BQ via Python: https://sophiamyang.github.io/DS/dataaccess/bigquery/bigquery.html
* Query via Python: https://cloud.google.com/bigquery/docs/quickstarts/quickstart-client-libraries?hl=de#run_the_query

* Query GBQ with pandas.gbq: https://stackoverflow.com/questions/44647310/convert-bigquery-results-to-pandas-data-frame
* Doc pandas-gbq: https://pandas-gbq.readthedocs.io/en/latest/install.html#conda
* Google doc zu pandas-gbq: https://cloud.google.com/bigquery/docs/pandas-gbq-migration?hl=de#conda
* Running Parametrized queries: https://cloud.google.com/bigquery/docs/parameterized-queries
* .to_dataframe doc: https://googleapis.dev/python/bigquery/latest/generated/google.cloud.bigquery.job.QueryJob.html#google.cloud.bigquery.job.QueryJob.to_dataframe
* Get BQ table as dataframe: https://cloud.google.com/bigquery/docs/samples/bigquery-list-rows-dataframe
* globals().clear()


# Packages

import pandas as pd
import numpy as np
from google.cloud import bigquery # when buggy: 'conda remove protobuf'
import os
import pyarrow


print(bigquery.__version__)

# Set env variable for credentials

os.environ["GOOGLE_APPLICATION_CREDENTIALS"]="/Users/marcvendramet/Desktop/HSG/Master/MA/08_Data/03_BQKey/mevtest-32e3980294f1.json"

client = bigquery.Client()

# Build Query Masterset

## Define query function

my_query = '''
SELECT *


FROM `mevtest.MEVextract.UniswapV2_transactions_download_v4`
WHERE block_number BETWEEN  @lower_bound AND @upper_bound 

ORDER BY block_number ASC, transaction_index ASC, log_index ASC

'''

## Collect results

bqCols = ['transaction_hash','block_number','transaction_index','log_index','from_address','to_address',
                            'trans_fromAddress','trans_toAddress', 'poolId', 'tf_tokenAddress', 'tokenName',
                            'value','trans_gasPrice','gasUsed','effective_gas_price']

dfMain = pd.DataFrame(columns = bqCols)

# Collect loop

## Collect manual

counter = 1
lowerB = 10093070
upperB = lowerB + 100000


job_config1 = bigquery.QueryJobConfig(
query_parameters=[
        bigquery.ScalarQueryParameter("lower_bound", "INT64", lowerB),
        bigquery.ScalarQueryParameter("upper_bound", "INT64", upperB),
    ]
    )
    
query_job = client.query(my_query, job_config=job_config1)
    
dfTemp = query_job.to_dataframe()
# create path variable
# save to.csv or parquet on ssd OK
dfTemp.to_parquet(f"/Volumes/Extreme SSD/02_UniswapV2Transactions/UniswapV2Transactions_{counter}.par")
# collect params from dfTemp andd update counter for new query OK
counter = counter + 1
lastEntry = dfTemp["block_number"].max().item() # OR: dfMain["block_number"].iloc[-1]
lowerB = lastEntry+1
upperB = lastEntry+100001

len(dfTemp.iloc[:,0]) >=1
print(lowerB, upperB, counter)

dfTemp.to_csv(f"/Volumes/Extreme SSD/02_UniswapV2Transactions/UniswapV2Transactions_2.csv")

dfTemp1 = dfTemp

lastEntry= lastEntry.item()
lowerB=lowerB.item()
upperB=upperB.item()

type(lastEntry.item())

## Automatic loop (full)

while True:
    # len(dfTemp.iloc[:,0]) >=1
    print(lowerB, upperB)
    job_config1 = bigquery.QueryJobConfig(
    query_parameters=[
        bigquery.ScalarQueryParameter("lower_bound", "INT64", lowerB),
        bigquery.ScalarQueryParameter("upper_bound", "INT64", upperB),
    ]
    )
    
    query_job = client.query(my_query, job_config=job_config1)
    
    dfTemp = query_job.to_dataframe()
    
    if dfTemp["block_number"].max() >= 12500188:
        dfTemp.to_parquet(f"/Volumes/Extreme SSD/02_UniswapV2Transactions/UniswapV2Transactions_{counter}.par")
        break
    else:
        # create path variable
        # save to.csv or parquet on ssd OK
        dfTemp.to_parquet(f"/Volumes/Extreme SSD/02_UniswapV2Transactions/UniswapV2Transactions_{counter}.par")
        # collect params from dfTemp andd update counter for new query OK
        counter = counter + 1
        lastEntry =dfTemp["block_number"].max().item() # OR: dfMain["block_number"].iloc[-1]
        lowerB = lastEntry+1
        upperB = lastEntry+100001
        print(lastEntry)

# Add nice print outs of status, loop number, part of loop its in, etc

dfTemp['block_number'].max()

dfTemp.shape

# Query: GasPrice Threeshold data set

my_query_GP = '''
SELECT *  
FROM `mevtest.MEVextract.GasPrice_threshold4`

WHERE block_number BETWEEN  11296630 AND 12500188 

'''

11296630-10093070


query_job = client.query(my_query_GP)
    





dfTemp = query_job.to_dataframe()


dfTemp['block_number'].max()

dfTemp.to_parquet(f"/Volumes/Extreme SSD/01_GasData/GasPriceThreshold_02.par")

a = (12500188-10093070)/2
10093070 + a

# Archive
## Query function
def query_mev():
    #client = bigquery.Client()
    query_job = client.query(
    """
    SELECT * 
    FROM `mevtest.MEVextract.LPs_UniswapV2`
    """)
    #results = query_job.result()
    #yield from results

## Potential eeasy way to store as pandas df
dfTest = query_job.to_dataframe() https://googleapis.dev/python/bigquery/latest/generated/google.cloud.bigquery.job.QueryJob.html#google.cloud.bigquery.job.QueryJob.to_dataframe

## unpack function def (used before .to_dataframe())
def query_unpack():
    results = query_job.result()
    yield from results

## Collect results and add to df
results2 = query_unpack()
dftemp = []
for row in results2:
    dftemp.append([
        row.block_number
    ])

dftemp = pd.DataFrame(dftemp, columns = ['block_number'])
dfMain = dfMain.append(dftemp)

## Query loop old version
while True:
    job_config1 = bigquery.QueryJobConfig(
    query_parameters=[
        bigquery.ScalarQueryParameter("lower_bound", "INT64", lowerB),
        bigquery.ScalarQueryParameter("upper_bound", "INT64", upperB),
    ]
    )
    
    query_job = client.query(my_query, job_config=job_config1)
    
    dfTemp = query_job.to_dataframe()
    
    if len(dfTemp.iloc[:,0]) <=1:
        break
    else:
        # create path variable
        # save to.csv or parquet on ssd OK
        dfTemp.to_parquet(f"/Volumes/Extreme SSD/02_UniswapV2Transactions/UniswapV2Transactions_{counter}.par")
        # collect params from dfTemp andd update counter for new query OK
        counter = counter + 1
        lastEntry =dfTemp["block_number"].max() # OR: dfMain["block_number"].iloc[-1]
        lowerB = lastEntry+1
        upperB = lastEntry+100001
        # check for max block
        if dfTemp["block_number"].max() >= 12500188:
            break
        else:
            # clean dfTemp
            dfTemp1 =dfTemp
            dfTemp =pd.DataFrame()
