==> 05_GasData_Preeparation_v2.ipynb <==
# Packages

import pandas as pd
import numpy as np
import math
pd.set_option('display.max_columns', None)
pd.set_option('max_colwidth', None) # show full width of showing cols

# Import file ThresholdGas v3

dfThreshold = pd.read_csv("/Volumes/Extreme SSD/01_GasData/GasPriceThreshold_v3.csv")

dfThreshold = dfThreshold.sort_values(by=['block_number', 'gas_price'])

dfThreshold = dfThreshold.drop_duplicates(subset = ["block_number"],keep = "first")

dfThreshold = dfThreshold.rename(columns={"gas_price": "threshold_gas"})

dfThreshold = dfThreshold.reset_index(drop=True)

# Import gasUseed per block from old GasFeePerBlock_BQ_v1

dfGasUsedBlock = pd.read_csv("/Volumes/Extreme SSD/01_GasData/GasFeePerBlock_BQ_v1.csv")

dfGasUsedBlock = dfGasUsedBlock.sort_values(by=['block_number'])

dfGasUsedBlock

# Converting to Gwei and add as new column

dfThreshold['threshold_gas']= (dfThreshold['threshold_gas']/(10**9)).astype(float)

dfThreshold

# Checks
* Start block: 19.5.2020 (go-live of UniswapV2) = 10,093,070 height
* End block: 24.5.2021 (V3 > V2) = 12,500,188 height


startB = 10093070
endB = 12500188
goalLength = endB-startB
goalLength

blockCount = dfThreshold['block_number'].max() - dfThreshold['block_number'].min()
blockCount
# 2407118

blockCount == goalLength

# Import Barbon file

dfBarb = pd.read_parquet("/Volumes/Extreme SSD/01_GasData/gas_ts_block.par")

dfBarb['block_number_bar'] = dfBarb.index

dfBarb.sort_values(by = "avg_gas", ascending = False)

# filter for thesis scope
#dfBarb_scope = dfBarb.filter(items = [*range(startB, endB+1)], axis=0) # Was machen wir hier?
dfBarb_scope = dfBarb[(dfBarb.block_number_bar >= startB) & (dfBarb.block_number_bar <= endB)]

#Check
dfBarb_scope['block_number_bar'].nunique() == len(dfBarb_scope.index) # check

# Missing blocks in barbon set: more detail in v1
dfBarb_scope['block_number_bar'].nunique() - goalLength

dfBarb_scope.sort_values(by = "avg_gas", ascending = False)

# Merge Barbin and Threshold files

# merge files
dfMerged = dfBarb_scope.merge(dfThreshold[['block_number','threshold_gas']], left_on = "block_number_bar", right_on='block_number')


dfMerged

dfMerged.info()

# Merge with total gas used per block from BQ

# merge files
dfMerged = dfMerged.merge(dfGasUsedBlock[['block_number','sumGasUsed']], on='block_number', how = 'left')


dfMerged = dfMerged.rename(columns={"sumGasUsed": "total_gasUsed"})

dfMerged.info()

dfMerged

# Save gasPriceAll-file

dfMerged.to_parquet("/Volumes/Extreme SSD/01_GasData/GasPriceAll.par")

dfMerged.to_csv("/Volumes/Extreme SSD/01_GasData/GasPriceAll.csv")
